requirements.txt
boto3==1.40.20
botocore==1.40.20
et_xmlfile==2.0.0
jmespath==1.0.1
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
s3transfer==0.13.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
